Updated to properly use async/await
We have not been making an actual HTTP call for configuration
requests. That was added now. However this required refactoring
a bit to properly support async requests.

Initializing the EppoClient now requires a load() function to be
called which actually retrieves data from the server and
makes it available to the app.